# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Alumni Spotlight API
  version: 1.0.0
  description: |
    API for alumni bidding system (Featured Alumni of the Day). Supports user auth, profile management, bidding, and admin workflows.
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
  /health:
    get:
      summary: Health check endpoint
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /auth/register:
    post:
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: User created
        "400":
          description: Validation error
  /auth/login:
    post:
      summary: Login and receive a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: JWT token
        "401":
          description: Invalid credentials
  /bids/status:
    get:
      summary: Check current bid status for authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current bid status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [winning, losing, none]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT